Moving classes to crud
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@5042 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Cache/CovCatC.cs
@@ -38,7 +38,7 @@ namespace OpenDentBusiness {
 		public static int GetOrderLong(long covCatNum) {
 			for(int i=0;i<Listt.Count;i++) {
 				if(covCatNum==Listt[i].CovCatNum) {
-					return i;
+					return (byte)i;
 				}
 			}
 			return -1;
